+++ crates/icydb-core/src/db/predicate/runtime/tests.rs
@@ -18,7 +18,7 @@ use std::borrow::Cow;
 
 static PREDICATE_FIELDS: [FieldModel; 4] = [
     FieldModel::generated("id", FieldKind::Ulid),
-    FieldModel::generated("score", FieldKind::Int),
+    FieldModel::generated("score", FieldKind::Int64),
     FieldModel::generated("tags", FieldKind::List(&FieldKind::Text { max_len: None })),
     FieldModel::generated("name", FieldKind::Text { max_len: None }),
 ];
