+++ crates/icydb-core/src/db/query/plan/tests/group.rs
@@ -52,7 +52,7 @@ crate::test_entity! {
     fields = [
         ("id", FieldKind::Ulid),
         ("tag", FieldKind::Text { max_len: None }),
-        ("rank", FieldKind::Int),
+        ("rank", FieldKind::Int64),
     ],
     indexes = [&INDEX_MODEL],
 }
