+++ crates/icydb-core/src/db/query/plan/tests/semantics.rs
@@ -63,7 +63,7 @@ crate::test_entity! {
     fields = [
         ("id", FieldKind::Ulid),
         ("tag", FieldKind::Text { max_len: None }),
-        ("rank", FieldKind::Int),
+        ("rank", FieldKind::Int64),
     ],
     indexes = [&INDEX_MODEL],
 }
@@ -115,7 +115,7 @@ crate::test_entity! {
     fields = [
         ("id", FieldKind::Ulid),
         ("name", FieldKind::Text { max_len: None }),
-        ("rank", FieldKind::Int),
+        ("rank", FieldKind::Int64),
     ],
     indexes = [&EXPRESSION_INDEX_MODEL],
 }
